==> 01_cnpj_empresas.ipynb <==
<B>TCC PUC MINAS - MAIO/2022</B><BR>
ANÁLISE DE PARCELAMENTO DE DÉBITOS INSCRITOS EM DÍVIDA ATIVA DA UNIÃO ATRAVÉS DE MACHINE LEARNING<BR>
FELIPE MARTINS<BR>

TRATAMENTO DOS ARQUIVOS DA BASE CNPJ / EMPRESA (8 DÍGITOS)

import pandas as pd

df0 = pd.read_csv("CNPJ_empresa/K3241.K03200Y0.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df0.head()

df0.dtypes

df1 = pd.read_csv("CNPJ_empresa/K3241.K03200Y1.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df2 = pd.read_csv("CNPJ_empresa/K3241.K03200Y2.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df3 = pd.read_csv("CNPJ_empresa/K3241.K03200Y3.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df4 = pd.read_csv("CNPJ_empresa/K3241.K03200Y4.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df5 = pd.read_csv("CNPJ_empresa/K3241.K03200Y5.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df6 = pd.read_csv("CNPJ_empresa/K3241.K03200Y6.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df7 = pd.read_csv("CNPJ_empresa/K3241.K03200Y7.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df8 = pd.read_csv("CNPJ_empresa/K3241.K03200Y8.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

df9 = pd.read_csv("CNPJ_empresa/K3241.K03200Y9.D10510.EMPRECSV.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 2, 5], names=['CNPJ_BASICO', 'NAT_JURID', 'PORTE'], dtype={'CNPJ_BASICO':object, 'NAT_JURID':object, 'PORTE':object})   #nrows=5

# CONCATENANDO OS 10 DATASETS GERADOS ACIMA
df_princ = pd.concat([df0, df1, df2, df3, df4, df5, df6, df7, df8, df9])

df_princ.count()

df_princ.dtypes

df_princ.count()

# EXPORTANTO ARQUIVO CONCATENADO
df_princ.to_csv("01_cnpj_empresa.csv")
