==> 02_cnpj_estabelecimentos.ipynb <==
<B>TCC PUC MINAS - MAIO/2022</B><BR>
ANÁLISE DE PARCELAMENTO DE DÉBITOS INSCRITOS EM DÍVIDA ATIVA DA UNIÃO ATRAVÉS DE MACHINE LEARNING<BR>
FELIPE MARTINS<BR>

TRATAMENTO DOS ARQUIVOS DA BASE CNPJ / ESTABELECIMENTOS (14 DÍGITOS)

import pandas as pd

# https://medium.com/@henrique.gelatti/diferentes-formas-de-ler-um-arquivo-csv-utilizando-a-biblioteca-pandas-dbeab96555ba
# https://pbpython.com/pandas_dtypes.html

df0 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y0.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df0.head()

df1 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y1.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df1.head()

df2 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y2.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df2.head()

df3 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y3.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df3.head()

df4 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y4.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df4.head()

df5 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y5.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df5.head()

df6 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y6.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df6.head()

df7 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y7.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df7.head()

df8 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y8.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df8.head()

df9 = pd.read_csv("CNPJ_estabelecimento/K3241.K03200Y9.D20212.ESTABELE.zip", sep= ';', encoding='ISO-8859-1', usecols=[0, 1, 2, 4, 5, 10, 11, 19], names=['CNPJ_BASICO', 'CNPJ_ORDEM', 'CNPJ_DV', 'NOME_FANTASIA', 'SIT_CADAST','DATA_INICIO_ATIV','CNAE_PRINC','UF'], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category"})   #nrows=5
df9.head()

# CONCATENANDO OS 10 DATASETS GERADOS ACIMA
df_cnpj_estab = pd.concat([df0, df1, df2, df3, df4, df5, df6, df7, df8, df9])

df_cnpj_estab.head()

df_cnpj_estab.count()

df_cnpj_estab.dtypes

# EXPORTANTO ARQUIVO CONCATENADO
df_cnpj_estab.to_csv("02_cnpj_estabelecimento.csv")
