==> 03_merge_empresa_estabelecimento.ipynb <==
<B>TCC PUC MINAS - MAIO/2022</B><BR>
ANÁLISE DE PARCELAMENTO DE DÉBITOS INSCRITOS EM DÍVIDA ATIVA DA UNIÃO ATRAVÉS DE MACHINE LEARNING<BR>
FELIPE MARTINS<BR>

FUSÃO DOS ARQUIVOS DA BASE CNPJ: EMPRESA + ESTABELECIMENTOS

import pandas as pd

# IMPORTANDO ARQUIVO 01_cnpj_empresa_completo.csv
df_empresa = pd.read_csv("01_cnpj_empresa.csv", usecols=[1, 2, 3], dtype={"CNPJ_BASICO":"category","NAT_JURID":"category","PORTE":"category"}) #sep= ';', encoding='ISO-8859-1')   #nrows=5
df_empresa.head()

# IMPORTANDO ARQUIVO 02_cnpj_estabelecimento.csv
df_estab = pd.read_csv("02_cnpj_estabelecimento.csv", usecols=[1, 2, 3, 5, 6, 7], dtype={"CNPJ_BASICO":"category","CNPJ_ORDEM":"category","CNPJ_DV":"category"}) #sep= ';', encoding='ISO-8859-1')   #nrows=5
df_estab.head()

# FUSÃO DAS 2 BASES DE CNPJ (ESTABELECIMENTO E EMPRESAS) COM OS CAMPOS SELECIONADOS (demorou 22 minutos de processamento)
# possível erro por insuficiência de memória ram. Neste caso, ver link abaixo:
# https://stackoverflow.com/questions/57507832/unable-to-allocate-array-with-shape-and-data-type

df_cnpj_merge = pd.merge(df_estab, df_empresa, how = 'left', on = 'CNPJ_BASICO')

df_cnpj_merge.head(10)

df_cnpj_merge.tail(10)

df_cnpj_merge.count()

# EXPORTANTO ARQUIVO APÓS FUSÃO (MERGE) BRUTO - ANTES DE SER TRABALHADO/AJUSTADP
df_cnpj_merge.to_csv("03_merge_empresa_estabelecimento.csv")
