==> 04_cnpj_final.ipynb <==
<B>TCC PUC MINAS - MAIO/2022</B><BR>
ANÁLISE DE PARCELAMENTO DE DÉBITOS INSCRITOS EM DÍVIDA ATIVA DA UNIÃO ATRAVÉS DE MACHINE LEARNING<BR>
FELIPE MARTINS<BR>

04 - CNPJ FINAL: AJUSTE
    
Tempo médio de processamento deste notebook: 50 minutos

import pandas as pd

%%time
df_cnpj_final = pd.read_csv("03_merge_empresa_estabelecimento.csv", usecols=[1, 2, 3, 4, 5, 6, 7, 8], dtype={"CNPJ_BASICO":"category", "CNPJ_ORDEM":"category", "CNPJ_DV":"category", "SIT_CADAST":"category", "DATA_INICIO_ATIV":"category", "CNAE_PRINC":"category", "NAT_JURID":"category", "PORTE":"category"}) #nrows=5
df_cnpj_final.head()

df_cnpj_final.count()

df_cnpj_final.isnull().sum()

df_cnpj_final.dtypes

df_cnpj_final.nunique()

%%time
# AJUSTANDO CNAE_PRINC
# O número original completo possui 7 dígitos e 1343 valores distintos.
# Analisando-se a estrutura do código, observa-se que 2 dígitos já são suficientes para identificar satisfatoriamente a atividade econômica.
df_cnpj_final['CNAE_PRINC'] = df_cnpj_final['CNAE_PRINC'].str[:2]

df_cnpj_final.head(15)

# CONVERTENDO DATA PARA STRING
df_cnpj_final["DATA_INICIO_ATIV"] = df_cnpj_final["DATA_INICIO_ATIV"].astype(str)

%%time
# CONVERTENDO CAMPO DATA_INICIO_ATIV PARA DATA
# coerce: will force the dates which are outside the bounds to 'NaT'. Há 5 registros com esse problema (devem ser datas muito antigas).

df_cnpj_final['DATA_INICIO_ATIV'] = pd.to_datetime(df_cnpj_final['DATA_INICIO_ATIV'], format="%Y/%m/%d", errors = 'coerce')  

df_cnpj_final.head()

df_cnpj_final.dtypes

%%time
# CRIANDO NOVO CAMPO COM ANO DO INÍCIO DA ATIVIDADE
df_cnpj_final['ANO_INICIO_ATIV'] = df_cnpj_final['DATA_INICIO_ATIV'].dt.year

df_cnpj_final.head()

%%time
# CRIANDO NOVO CAMPO COM ANO ATIVIDADE MAIOR DO QUE 10 ANOS (contados do ano da base PGFN, que é 2021)
# não consegui fazer com pandas (acima). Mas deu certo com numpy:
import numpy as np
df_cnpj_final['ATIV_>10_ANOS'] = np.where(df_cnpj_final.ANO_INICIO_ATIV<=2010, 'S', 'N')

df_cnpj_final.head()

df_cnpj_final.tail()

print(df_cnpj_final["ATIV_>10_ANOS"].value_counts())

df_cnpj_final.dtypes

# O MERGE NÃO DEU CERTO PARA 3 milhões de REGISTROS. NÃO SEI SE FOI ALGUM ERRO OU SE REALMENTE NÃO EXISTE CORRESPONDENTE.
df_cnpj_final.isnull().sum()

# EXCLUINDO COLUNAS DESNECESSÁRIAS (para liberar memória)
df_cnpj_final = df_cnpj_final.drop(columns=['DATA_INICIO_ATIV','ANO_INICIO_ATIV'])

%%time
# CRIANDO NOVO CAMPO CNPJ COM PONTUAÇÃO (SE DER PROBLEMA DE MEMÓRIA, FECHAR OUTROS NOTEBOOKS ABERTOS)
df_cnpj_final['CNPJ_BASICO_2'] = df_cnpj_final['CNPJ_BASICO'].str[0:2]+"."+df_cnpj_final['CNPJ_BASICO'].str[2:5]+"."+df_cnpj_final['CNPJ_BASICO'].str[5:9]

df_cnpj_final.head()

%%time
# UNINDO AS 3 COLUNAS CNPJ
df_cnpj_final['CNPJ_BASICO_3'] = df_cnpj_final['CNPJ_BASICO_2'].str[0:]+"/"+df_cnpj_final['CNPJ_ORDEM'].str[0:]+"-"+df_cnpj_final['CNPJ_DV'].str[0:]

df_cnpj_final.head()

# EXCLUINDO COLUNAS DESNECESSÁRIAS
df_cnpj_final = df_cnpj_final.drop(columns=['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','CNPJ_BASICO_2'])

df_cnpj_final.head()

#REORDENANDO COLUNAS
df_cnpj_final = df_cnpj_final[['CNPJ_BASICO_3','SIT_CADAST','CNAE_PRINC','NAT_JURID','PORTE','ATIV_>10_ANOS']]

df_cnpj_final.head()

%%time
# RENOMEANDO COLUNA
df_cnpj_final = df_cnpj_final.rename(columns = {'CNPJ_BASICO_3': 'CNPJ_BASICO'})

%%time
# EXPORTANTO DATASET FINAL
df_cnpj_final.to_csv("04_cnpj_final.csv")

df_cnpj_final.info()
