==> 05_debitos.ipynb <==
<B>TCC PUC MINAS - MAIO/2022</B><BR>
ANÁLISE DE PARCELAMENTO DE DÉBITOS INSCRITOS EM DÍVIDA ATIVA DA UNIÃO ATRAVÉS DE MACHINE LEARNING<BR>
FELIPE MARTINS<BR>

05 - DÉBITOS PGFN
    
IMPORTAÇÃO E TRATAMENTO DO DATASET DO ESTOQUE DE DÉBITOS INSCRITOS EM DAU EM 1TRIM/2022
    
Tempo médio de processamento completo deste notebook: 10 min

import pandas as pd

## DATASET DE DEBITOS DA PGFN DO 1º TRIMESTRE/2022 (df) ##

%%time
# IMPORTANDO DATASET DE DEBITOS DA PGFN DO 4º TRIMESTRE/2020
# Importando dataset de uma pasta compactada com vários arquivos e concatenando-os
# https://samukweku.github.io/data-wrangling-blog/python/pandas/compressed%20data/zip/archived%20data/2020/07/21/Extract-DataFrame-from-Compressed-Data-into-Pandas.html

from zipfile import ZipFile

with ZipFile("pgfn\Dados_abertos_Nao_Previdenciario_1trim2022.zip") as zipfiles:
    #the first entry is the zipfile name
    #we'll skip it
    filelist = zipfiles.namelist()[1:]
    
    data = [pd.read_csv(zipfiles.open(file_name), usecols=[0, 1, 2, 6, 7, 8, 9, 10, 11, 12], dtype={"NUMERO_INSCRICAO":"category"}, sep= ';', encoding='unicode_escape') for file_name in filelist]  #nrows=5

df = pd.concat(data)

df.head()

# RESETANDO O INDEX
# MOTIVO: AO IMPORTAR VÁRIOS ARQUIVOS AO MESMO TEMPO, O INDICE REINICIA DO ZERO PARA CADA ARQUIVO.
# POR EXEMPLO: COMO SÃO 26 ARQUIVOS, O DATASET FICA COM 26 LINHAS DE INDICES IGUAIS (26 LINHAS INDICE 0, 26 LINHAS INDICE 1, ...)
# ISSO COMPROMETE TODA A MANIPULAÇÃO DO DS, POIS AO SE TENTAR ALTERAR UMA LINHA ESPECÍFICA, TODAS AS LINHAS DE MESMO INDICE SÃO ALTERADAS.
# O COMANDO ABAIXO RESETA O INDICE E RESOLVE O PROBLEMA. O PARÂMETRO drop=True É PARA NÃO CRIAR UMA NOVA COLUNA COM ÍNDICE.
df = df.reset_index(drop=True)

df.head()

df.count()

df.dtypes

%%time
# CONVERTENDO CAMPO DATA_INSCRICAO PARA TIPO DATETIME
df["DATA_INSCRICAO"] = df["DATA_INSCRICAO"].astype("datetime64")

df.dtypes

df.head()

df.isnull().sum()

## Ajustando campo TIPO_DEVEDOR
Removendo-se o tipo de devedor diferente de principal

print(df["TIPO_DEVEDOR"].value_counts())

%%time
# REMOVENDO TODOS AS LINHAS DA COLUNA 'TIPO_DEVEDOR' QUE SÃO DIFERENTES DE 'PRINCIPAL' (OU SEJA, REMOVENDO SOLIDÁRIOS E CORRESPONSÁVEL PARA EVITAR DUPLICIDADE DE INSCRIÇÃO)
# PROCESSAMENTO DEMORADO (cerca de 10min)
df = df.drop(df.loc[df['TIPO_DEVEDOR']!='PRINCIPAL'].index)

print(df["TIPO_DEVEDOR"].value_counts())

df.count()

df.head()

## Ajustando campo TIPO_PESSOA
Substituindo-se os valores do campo ‘TIPO_PESSOA’ por PF e PJ

%%time
# SUBSTITUINDO O CAMPO TIPO_PESSOA POR PJ E PF
df['TIPO_PESSOA'] = df['TIPO_PESSOA'].replace(['Pessoa jurídica','Pessoa física'],['PJ','PF'])

print(df["TIPO_PESSOA"].value_counts())

df.head()

df.count()

df.dtypes

## Criação do campo INSC_>1_ANO

%%time
# CRIANDO NOVO CAMPO CALCULADO DO TEMPO DE INSCRIÇÃO DA DÍVIDA (MAIOR QUE 1 ANO) UTILIZANDO DATA_INSCRICAO
# ATENÇÃO: PERÍODO-BASE É 1 TRIM/2022. ASSIM, SERÁ CONSIDERADA DÍVIDA MAIOR QUE 1 ANO AS INSCRITAS ATÉ 2019.
import numpy as np

df['ANO_INSCRICAO'] = df['DATA_INSCRICAO'].dt.year
df['INSC_>1_ANO'] = np.where(df.ANO_INSCRICAO<=2019, 'S', 'N')

df.head(100)

## Criação do campo HA_OUTROS_DEBITOS

%%time
# CRIANDO CAMPO QUE VERIFICA SE HÁ MAIS DE UM DÉBITO (MAIS DE UMA LINHA PARA O MESMO CPF_CNPJ).

import numpy as np
df['CONTADOR_NOME'] = df.groupby('CPF_CNPJ')['CPF_CNPJ'].transform('count')
df['HA_OUTROS_DEBITOS'] = np.where(df.CONTADOR_NOME>1, 'S', 'N')

df.head()

## Criação do campo FAIXA DE VALOR

%%time
# CRIANDO NOVO CAMPO CALCULADO DA FAIXA DE VALOR (https://insightlab.ufc.br/6-truques-do-pandas-para-impulsionar-sua-analise-de-dados/ - VER ITEM 4)

import sys

 
df['VALOR_FAIXA']=pd.cut(

   df['VALOR_CONSOLIDADO'],

   bins=[0, 1000, 5000, 10000, 20000, 50000, 100000, 500000, 1000000, 10000000, sys.maxsize],

   labels=['1', '2', '3', '4', '5', '6', '7', '8', '9', '10']
)

# https://re-thought.com/pandas-value_counts/
print(df["VALOR_FAIXA"].value_counts().sort_index(ascending=True))

## Limpando o campo 'RECEITA_PRINCIPAL'

# ANTES DA LIMPEZA DA COLUNA RECEITA_PRINCIPAL
print(df["RECEITA_PRINCIPAL"].value_counts())

%%time
# LIMPANDO A COLUNA RECEITA PRINCIPAL (REMOVER TEXTO 'Receita da dívida ativa' e afins)
#https://pt.stackoverflow.com/questions/491195/em-python-como-remover-caracteres-espec%C3%ADficos-de-todos-os-registros-de-apenas-u
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('Receita da dívida ativa - ', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace(' - Receita da dívida ativa', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('Rec. dív. ativa - ', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('Rec. dívida ativa - ', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('Receita da dív. ativa - ', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('DIV.ATIVA-', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('DIV ATIVA-', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('R D ATIVA - ', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('R D Ativa - ', '')
df["RECEITA_PRINCIPAL"] = df["RECEITA_PRINCIPAL"].str.replace('R D Ativa -', '')

# DEPOIS DA LIMPEZA
print(df["RECEITA_PRINCIPAL"].value_counts())

df.head(100)

## Reduzindo a quantidade de valores do campo 'RECEITA_PRINCIPAL'

# CONTAGEM DE VALORES RECEITA_PRINCIPAL
count_receita_princ = df['RECEITA_PRINCIPAL'].value_counts()
count_receita_princ.head()

%%time
# Redução de valores do campo 'RECEITA_PRINCIPAL'
# Os dados originais apresentam mais de 100 variedades de tributos, a maioria com baixa frequencia, conforme acima.
# Assim, escolheremos apenas os tributos mais frequentes para manterem seus nomes originais. Os demais serão renomeados para "OUTROS".
def replace(x):
    if x != "SIMPLES NACIONAL" and x != "IRPF" and x != "COFINS" and x != "IRPJ" and x != "CSLL" and x != "PIS" and x != "Multa - CLT" and x != "Multa Isolada" and x != "SIMPLES" and x != "SPU" and x != "IRRF" and x != "SIMP NAC - MEI":
        return 'OUTROS'
    else:
        return str(x)

df['RECEITA_PRINCIPAL'] = df['RECEITA_PRINCIPAL'].apply(replace)

## Criação do campo target PARCELADO

# IDENTIFICANDO VALOR QUE CORRESPONDE AOS DEBITOS PARCELADOS
print(df["TIPO_SITUACAO_INSCRICAO"].value_counts())

# AQUI SE OBSERVA QUE O TIPO DE SITUAÇÃO "BENEFÍCIO FISCAL" CONTEM OS DEBITOS QUE FORAM PARCELADOS (PARCELADA, NEGOCIADA NO SISPAR, REFIS, PAES, ETC.). 
df.groupby('TIPO_SITUACAO_INSCRICAO')['SITUACAO_INSCRICAO'].value_counts()

%%time
# CRIANDO NOVO CAMPO CALCULADO: PARCELADO (S/N)

import numpy as np

df['PARCELADO'] = np.where(df.TIPO_SITUACAO_INSCRICAO=='Benefício Fiscal', 'S', 'N')

# CONFERÊNCIA: OBSERVE QUE A QDE DE SIM CONFERE COM A QDE DE BENEFÍCIO FISCAL (3.452.760)
print(df["PARCELADO"].value_counts())

## Procedimentos finais

df.count()

df.head(100)

df.tail(100)

# EXCLUINDO COLUNAS DESNECESSÁRIAS
df = df.drop(columns=['DATA_INSCRICAO','ANO_INSCRICAO','CONTADOR_NOME', 'TIPO_DEVEDOR'])

df.info()

df.head()

%%time
# EXPORTANTO DATASET
df.to_csv("05_debitos.csv")
