==> 06_merge_debitos_cnpj.ipynb <==
<B>TCC PUC MINAS - MAIO/2022</B><BR>
ANÁLISE DE PARCELAMENTO DE DÉBITOS INSCRITOS EM DÍVIDA ATIVA DA UNIÃO ATRAVÉS DE MACHINE LEARNING<BR>
FELIPE MARTINS<BR>

06 - FUSÃO (MERGE) DOS DATASETS DEBITOS E CNPJ
    
Tempo médio de processamento deste notebook: 90 minutos.

import pandas as pd

## Importadando dataset da base CNPJ

%%time
# IMPORTANDO ARQUIVO 04_cnpj_final.csv
df_cnpj_final = pd.read_csv("04_cnpj_final.csv", dtype={"CNPJ_BASICO":"category", "NAT_JURID":"category", "PORTE":"category"})  #nrows=5

df_cnpj_final.head()

df_cnpj_final = df_cnpj_final.drop(columns=['Unnamed: 0'])

# RENOMEANDO COLUNA (APESAR DE O DATASET CNPJ NÃO TER CPF, É NECESSÁRIO O MESMO NOME DE COLUNA PARA O MERGE)
df_cnpj_final = df_cnpj_final.rename(columns = {'CNPJ_BASICO': 'CPF_CNPJ'})

df_cnpj_final.head()

df_cnpj_final.info()

## Importadando dataset de débitos

%%time
# IMPORTANDO ARQUIVO 05_debitos.csv
df_debitos = pd.read_csv("05_debitos.csv")   #nrows=5

df_debitos = df_debitos.drop(columns=['Unnamed: 0'])

df_debitos.head()

df_debitos.info()

## Realizando-se a fusão (merge) dos datasets

%%time
# FUSÃO DOS DATASET DÉBITOS E CNPJ_FINAL
# PROCESSAMENTO DEMORADO. POSSÍVEL PROBLEMA DE FALTA DE MEMÓRIA. ENCERRAR OUTROS PROGRAMAS E NOTEBOOKS ABERTOS.
df_debitos_cnpj = pd.merge(df_debitos, df_cnpj_final, how = 'left', on = 'CPF_CNPJ')

## Procedimentos finais

df_debitos_cnpj.head(100)

df_debitos_cnpj.info()

# AJUSTANDO TIPO DE DADOS RESTANTES PARA OBJECT/STRING
df_debitos_cnpj["NUMERO_INSCRICAO"] = df_debitos_cnpj["NUMERO_INSCRICAO"].astype("object")
df_debitos_cnpj["SIT_CADAST"] = df_debitos_cnpj["SIT_CADAST"].astype("object")
df_debitos_cnpj["CNAE_PRINC"] = df_debitos_cnpj["CNAE_PRINC"].astype("object")
df_debitos_cnpj["NAT_JURID"] = df_debitos_cnpj["NAT_JURID"].astype("object")
df_debitos_cnpj["PORTE"] = df_debitos_cnpj["PORTE"].astype("object")

df_debitos_cnpj.info()

%%time
# EXPORTANTO DATASET FINAL (antes da exclusão das colunas desnecessárias para ML)
# Essa linha de comando nao consta no TCC. Foram utilizados apenas para uso interno.
df_debitos_cnpj.to_csv("06_dataset_final_antes_limpeza.csv")

# EXCLUINDO COLUNAS DESNECESSÁRIAS (LIMPEZA FINAL)
df_debitos_cnpj = df_debitos_cnpj.drop(columns=['CPF_CNPJ','NUMERO_INSCRICAO'])

df_debitos_cnpj.head()

df_debitos_cnpj.tail()

df_debitos_cnpj.nunique()

# REORDENANDO COLUNAS (ATRIBUTO-CLASSE (ALVO) (PARCELADO) DEVE SER A ÚLTIMA)
df_debitos_cnpj = df_debitos_cnpj[['TIPO_PESSOA','TIPO_SITUACAO_INSCRICAO','SITUACAO_INSCRICAO','RECEITA_PRINCIPAL','INDICADOR_AJUIZADO','HA_OUTROS_DEBITOS','INSC_>1_ANO','VALOR_CONSOLIDADO','VALOR_FAIXA','SIT_CADAST','CNAE_PRINC','NAT_JURID','PORTE','ATIV_>10_ANOS','PARCELADO']]

df_debitos_cnpj.head(15)

%%time
# EXPORTANTO DATASET FINAL SEM CPF/CNPJ, NUMERO_INSCRICAO
# TODOS OS DADOS AINDA SÃO CATEGÓRICOS. SERÃO NECESSÁRIOS EM 08_analise_exploratoria.ypynb
# PARA MACHINE LEARNING, SERÁ NECESSÁRIA A CONVERSÃO PARA DADOS CONTÍNUOS. SERÁ FEITA EM 07_conv_dados_continuos.ipynb.
df_debitos_cnpj.to_csv("06_dataset_final_com_dados_categoricos.csv")

df_debitos_cnpj.info()
