==> 07_conv_dados_continuos.ipynb <==
<B>TCC PUC MINAS - MAIO/2022</B><BR>
ANÁLISE DE PARCELAMENTO DE DÉBITOS INSCRITOS EM DÍVIDA ATIVA DA UNIÃO ATRAVÉS DE MACHINE LEARNING<BR>
FELIPE MARTINS<BR>

07 - CONVERTENDO TIPO DE DADOS PARA DADOS QUANTITATIVOS
    
Tempo aproximado de processamento: 5 minutos

# 1. Prepraro inicial

import pandas as pd
from sklearn.model_selection import StratifiedShuffleSplit

%%time
# IMPORTANDO O DATASET
df = pd.read_csv("06_dataset_final_com_dados_categoricos.csv")

df = df.drop(columns=['Unnamed: 0'])

df.head()

df.info()

### Reduzindo a quantidade de linhas do dataset (redução de numerosidade / amostragem estratificada)

# Contando a quantidade de classes no dataset original
df['PARCELADO'].value_counts()

3452760 / len(df)  # proporção da classe minoritária. Após o split conferir se será mantida.

%%time
# Dividindo o dataset
split = StratifiedShuffleSplit(test_size=0.05)
for x, y in split.split(df, df['PARCELADO']):
    df1 = df.iloc[x]  # 95% dos dados
    df2 = df.iloc[y]  # 5% dos dados. Utilizar essa.

# Resultado da divisão
print("dataset original: ",df.shape)
print("dataset 95%: ",df1.shape)
print("dataset 5% - nossa amostra: ",df2.shape)

# Contando a quantidade de classes no dataset reduzido
df2['PARCELADO'].value_counts()

172638 / len(df2)  # confirmação de que proporção da classe minoritária foi mantida.

df2

### Verificação do dataset reduzido

# CHECANDO SE HÁ DADOS NULOS / NAN (NOT A NUMBER)
# RESULTADO: COLUNAS VINDAS DA BASE CNPJ POSSUEM MAIS DE 2 MILHÕES DE DADOS NULOS. SE REFEREM AOS DEBITOS DE PESSOA FÍSICA, QUE, OBVIAMENTE, NÃO TÊM REGISTRO NA BASE CNPJ.
df2.isna().sum()

df2['TIPO_PESSOA'].value_counts()  #contador de frequência

# SUBSTITUINDO DADOS NULOS / NAN POR ZERO PARA EVITAR ERROS NAS CONVERSÕES
df2.fillna(value = 0, inplace = True)
df2.isna().sum()

# ENUMERA A QDE DE VARIAÇÕES DE CADA COLUNA
df2.nunique()

df2.dtypes

## PARTE 2 - NORMALIZANDO VARIÁVEIS CONTÍNUAS
MÉTODO: MinMaxScaler

%%time
from sklearn.preprocessing import MinMaxScaler
variaveis_continuas = ['VALOR_FAIXA', 'VALOR_CONSOLIDADO']
scaler = MinMaxScaler()
df2[variaveis_continuas] = scaler.fit_transform(df2[variaveis_continuas])
df2[variaveis_continuas]

## PARTE 3 - CONVERTENDO VARIÁVEIS CATEGÓRICAS BINÁRIAS
MÉTODO: REPLACE

%%time
df2['TIPO_PESSOA'] = df2['TIPO_PESSOA'].replace(['PJ','PF'],['1','0'])
df2['INDICADOR_AJUIZADO'] = df2['INDICADOR_AJUIZADO'].replace(['SIM','NAO'],['1','0'])
df2['HA_OUTROS_DEBITOS'] = df2['HA_OUTROS_DEBITOS'].replace(['S','N'],['1','0'])
df2['INSC_>1_ANO'] = df2['INSC_>1_ANO'].replace(['S','N'],['1','0'])
df2['ATIV_>10_ANOS'] = df2['ATIV_>10_ANOS'].replace(['S','N'],['1','0'])
df2['PARCELADO'] = df2['PARCELADO'].replace(['S','N'],['1','0'])
df2.head()

%%time
# CONVERTENDO TIPO DE DADOS PARA INTEIRO
df2["TIPO_PESSOA"] = df2["TIPO_PESSOA"].astype("int64")
df2["INDICADOR_AJUIZADO"] = df2["INDICADOR_AJUIZADO"].astype("int64")
df2["HA_OUTROS_DEBITOS"] = df2["HA_OUTROS_DEBITOS"].astype("int64")
df2["INSC_>1_ANO"] = df2["INSC_>1_ANO"].astype("int64")
df2["ATIV_>10_ANOS"] = df2["ATIV_>10_ANOS"].astype("int64")
df2["PARCELADO"] = df2["PARCELADO"].astype("int64")

df2.info()

## PARTE 4 - CONVERTENDO  VARIÁVEIS CATEGÓRICAS MULTICLASSE
UTILIZAÇÃO DE ENCODER WOEEncoder

# CONVERTENDO A CATEGORIA-ALVO (PARCELADO) PARA INTEIRO, POIS É NECESSÁRIO PARA OS ENCODERS A SEGUIR
df2["PARCELADO"] = df2["PARCELADO"].astype("int64")

pip install category_encoders

%%time
from category_encoders import WOEEncoder

y = df2['PARCELADO']
X = df2['TIPO_SITUACAO_INSCRICAO']
encoder = WOEEncoder(cols=['TIPO_SITUACAO_INSCRICAO']).fit(X, y)
df2['COD_TIPO_SIT_INSC'] = encoder.transform(X)
df2[['TIPO_SITUACAO_INSCRICAO', 'COD_TIPO_SIT_INSC']].groupby('TIPO_SITUACAO_INSCRICAO').agg(lambda x: list(set(x))).reset_index()


%%time
from category_encoders import WOEEncoder

y = df2['PARCELADO']
X = df2['SITUACAO_INSCRICAO']
encoder = WOEEncoder(cols=['SITUACAO_INSCRICAO']).fit(X, y)
df2['COD_SIT_INSC'] = encoder.transform(X)
df2[['SITUACAO_INSCRICAO', 'COD_SIT_INSC']].groupby('SITUACAO_INSCRICAO').agg(lambda x: list(set(x))).reset_index()


%%time
y = df2['PARCELADO']
X = df2['RECEITA_PRINCIPAL']
encoder = WOEEncoder(cols=['RECEITA_PRINCIPAL']).fit(X, y)
df2['COD_RECEITA'] = encoder.transform(X)
df2[['RECEITA_PRINCIPAL', 'COD_RECEITA']].groupby('RECEITA_PRINCIPAL').agg(lambda x: list(set(x))).reset_index()


%%time
y = df2['PARCELADO']
X = df2['SIT_CADAST']
encoder = WOEEncoder(cols=['SIT_CADAST']).fit(X, y)
df2['COD_SIT_CADAST'] = encoder.transform(X)
df2[['SIT_CADAST', 'COD_SIT_CADAST']].groupby('SIT_CADAST').agg(lambda x: list(set(x))).reset_index()


%%time
y = df2['PARCELADO']
X = df2['CNAE_PRINC']
encoder = WOEEncoder(cols=['CNAE_PRINC']).fit(X, y)
df2['COD_CNAE_PRINC'] = encoder.transform(X)
df2[['CNAE_PRINC', 'COD_CNAE_PRINC']].groupby('CNAE_PRINC').agg(lambda x: list(set(x))).reset_index()

df2['COD_CNAE_PRINC'].value_counts()  #contador de frequência

%%time
y = df2['PARCELADO']
X = df2['NAT_JURID']
encoder = WOEEncoder(cols=['NAT_JURID']).fit(X, y)
df2['COD_NAT_JURID'] = encoder.transform(X)
df2[['NAT_JURID', 'COD_NAT_JURID']].groupby('NAT_JURID').agg(lambda x: list(set(x))).reset_index()

df2['COD_NAT_JURID'].value_counts()  #contador de frequência

%%time
y = df2['PARCELADO']
X = df2['PORTE']
encoder = WOEEncoder(cols=['PORTE']).fit(X, y)
df2['COD_PORTE'] = encoder.transform(X)
df2[['PORTE', 'COD_PORTE']].groupby('PORTE').agg(lambda x: list(set(x))).reset_index()


df2['COD_PORTE'].value_counts()  #contador de frequência

## PARTE 5 - PROCEDIMENTOS FINAIS

df2.head()

# EXCLUINDO COLUNAS REDUNDANTES
df2 = df2.drop(columns=['TIPO_SITUACAO_INSCRICAO','SITUACAO_INSCRICAO','RECEITA_PRINCIPAL','SIT_CADAST','CNAE_PRINC','NAT_JURID','PORTE'])

# REORDENANDO COLUNAS
df2 = df2[['TIPO_PESSOA','COD_TIPO_SIT_INSC','COD_SIT_INSC','COD_RECEITA','INDICADOR_AJUIZADO','HA_OUTROS_DEBITOS','INSC_>1_ANO','VALOR_CONSOLIDADO','VALOR_FAIXA','COD_SIT_CADAST','COD_CNAE_PRINC','COD_NAT_JURID','COD_PORTE','ATIV_>10_ANOS','PARCELADO']]
df2.head()

# CHECANDO MAIS UMA VEZ SE HÁ DADOS NULOS / NAN (NOT A NUMBER)
df2.isna().sum()

# SUBSTITUINDO DADOS NULOS / NAN POR ZERO
df2.fillna(value = 0, inplace = True)
df2.isna().sum()

%%time
# EXPORTANTO DATASET FINAL
df2.to_csv("07_dataset_final_com_dados_continuos.csv")

df2.info()
